Refactored screens package so subpackage names are consistent.
