+++ source/core/assets/configs/leaderboard.txt
@@ -26,3 +26,4 @@ DirtyDefault39718:958,
 DirtyDefault87630:947,
 DirtyDefault64325:942,
 DirtyDefault37738:849,
+DirtyDefault38752:780,
