+++ source/core/src/main/com/deco2800/game/input/KeyboardInputFactory.java
@@ -1,7 +1,7 @@
 package com.deco2800.game.input;
 import com.deco2800.game.entities.components.player.KeyboardPlayerInputComponent;
 import com.deco2800.game.input.components.InputComponent;
-import com.deco2800.game.screens.titlescreen.KeyboardTitleInputComponent;
+import com.deco2800.game.screens.title.KeyboardTitleInputComponent;
 import com.deco2800.game.ui.terminal.KeyboardTerminalInputComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
