+++ source/core/src/main/com/deco2800/game/screens/context/ContextScreenDisplay.java
@@ -1,4 +1,4 @@
-package com.deco2800.game.screens.contextscreen;
+package com.deco2800.game.screens.context;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
@@ -6,7 +6,7 @@ import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.deco2800.game.generic.ServiceLocator;
-import com.deco2800.game.screens.EndGameDisplay;
+import com.deco2800.game.screens.endgame.EndGameDisplay;
 import com.deco2800.game.ui.components.UIComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
