+++ source/core/src/main/com/deco2800/game/screens/endgame/EndGameScreen.java
@@ -1,42 +1,54 @@
-package com.deco2800.game.screens.contextscreen;
+package com.deco2800.game.screens.endgame;
 
 import com.badlogic.gdx.ScreenAdapter;
 import com.badlogic.gdx.scenes.scene2d.Stage;
+import com.deco2800.game.GdxGame;
 import com.deco2800.game.entities.Entity;
 import com.deco2800.game.entities.EntityService;
 import com.deco2800.game.entities.factories.RenderFactory;
-import com.deco2800.game.generic.ResourceService;
-import com.deco2800.game.generic.ServiceLocator;
-import com.deco2800.game.input.InputService;
 import com.deco2800.game.input.components.InputDecorator;
+import com.deco2800.game.input.InputService;
 import com.deco2800.game.rendering.RenderService;
 import com.deco2800.game.rendering.Renderer;
+import com.deco2800.game.generic.ResourceService;
+import com.deco2800.game.generic.ServiceLocator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * The game screen containing the context.
+ * The win/lose screen at the end of the game.
  */
-public class ContextScreen extends ScreenAdapter {
-    private static final Logger logger = LoggerFactory.getLogger(ContextScreen.class);
-    private static final String[] ContextTextures = {
-            "images/context_screen/context_screen.PNG",
-            "images/ui/screens/inactiveStart.png"
-    };
+public class EndGameScreen extends ScreenAdapter {
+    private static final Logger logger = LoggerFactory.getLogger(EndGameScreen.class);
+    private static final String[] winScreenTextures = {"images/ui/screens/win_screen.png"};
+    private static final String[] loseScreenTextures = {"images/ui/screens/lose_screen.png"};
+    private static final String[] timeoutScreenTextures = {"images/ui/screens/time_out.png"};
+    private final String[] activeScreenTextures;
+    private final GdxGame.ScreenType result;
 
     private final Renderer renderer;
 
-    public ContextScreen() {
+    public EndGameScreen(GdxGame.ScreenType result) {
+        this.result = result;
+        switch (this.result) {
+            case WIN_DEFAULT:
+                this.activeScreenTextures = winScreenTextures;
+                break;
+            case LOSS_TIMED:
+                this.activeScreenTextures = timeoutScreenTextures;
+                break;
+            case LOSS_CAUGHT:
+            default:
+                this.activeScreenTextures = loseScreenTextures;
+        }
 
-        logger.debug("Initialising Context screen services");
+        logger.debug("Initialising end game screen services");
         ServiceLocator.registerInputService(new InputService());
         ServiceLocator.registerResourceService(new ResourceService());
         ServiceLocator.registerEntityService(new EntityService());
         ServiceLocator.registerRenderService(new RenderService());
 
         renderer = RenderFactory.createRenderer();
-
-        System.out.println();
 
         loadAssets();
         createUI();
@@ -66,7 +78,7 @@ public class ContextScreen extends ScreenAdapter {
 
     @Override
     public void dispose() {
-        logger.debug("Disposing context screen");
+        logger.debug("Disposing end game screen");
 
         renderer.dispose();
         unloadAssets();
@@ -77,30 +89,38 @@ public class ContextScreen extends ScreenAdapter {
         ServiceLocator.clear();
     }
 
+    public GdxGame.ScreenType getResult() {
+        return this.result;
+    }
+
+    public String[] getActiveScreenTextures() {
+        return this.activeScreenTextures;
+    }
+
     private void loadAssets() {
         logger.debug("Loading assets");
         ResourceService resourceService = ServiceLocator.getResourceService();
-        resourceService.loadTextures(ContextTextures);
+        resourceService.loadTextures(activeScreenTextures);
         ServiceLocator.getResourceService().loadAll();
     }
 
     private void unloadAssets() {
         logger.debug("Unloading assets");
         ResourceService resourceService = ServiceLocator.getResourceService();
-        resourceService.unloadAssets(ContextTextures);
+        resourceService.unloadAssets(activeScreenTextures);
     }
 
     /**
-     * Creates the context screens ui including components for rendering ui elements to the screen and
+     * Creates the end game's ui including components for rendering ui elements to the screen and
      * capturing and handling ui input.
      */
     private void createUI() {
         logger.debug("Creating ui");
         Stage stage = ServiceLocator.getRenderService().getStage();
         Entity ui = new Entity();
-        ui.addComponent(new ContextScreenDisplay())
+        ui.addComponent(new EndGameDisplay(this))
                 .addComponent(new InputDecorator(stage, 10))
-                .addComponent(new ContextScreenActions());
+                .addComponent(new EndGameActions());
         ServiceLocator.getEntityService().register(ui);
     }
 }
