+++ source/core/src/test/com/deco2800/game/screens/ContextScreenActionsTest.java
@@ -3,7 +3,7 @@ package com.deco2800.game.screens;
 import com.deco2800.game.GdxGame;
 import com.deco2800.game.extensions.GameExtension;
 import com.deco2800.game.generic.ServiceLocator;
-import com.deco2800.game.screens.contextscreen.ContextScreenActions;
+import com.deco2800.game.screens.context.ContextScreenActions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
