+++ source/core/src/test/com/deco2800/game/screens/EndGameActionsTest.java
@@ -3,10 +3,9 @@ package com.deco2800.game.screens;
 import com.deco2800.game.GdxGame;
 import com.deco2800.game.extensions.GameExtension;
 import com.deco2800.game.generic.ServiceLocator;
+import com.deco2800.game.screens.endgame.EndGameActions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.security.Provider;
 
 import static org.mockito.Mockito.*;
