+++ source/core/src/test/com/deco2800/game/screens/LeaderBoardDisplayTest.java
@@ -1,7 +1,7 @@
 package com.deco2800.game.screens.endgame;
 import com.deco2800.game.GdxGame;
 import com.deco2800.game.extensions.GameExtension;
-import com.deco2800.game.screens.leaderboardscreen.LeaderBoardDisplay;
+import com.deco2800.game.screens.leaderboard.LeaderBoardDisplay;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import static org.junit.jupiter.api.Assertions.*;
