+++ source/core/src/test/com/deco2800/game/screens/LeaderBoardScreenTest.java
@@ -2,8 +2,8 @@ package com.deco2800.game.screens.endgame;
 
 import com.deco2800.game.GdxGame;
 import com.deco2800.game.extensions.GameExtension;
-import com.deco2800.game.screens.leaderboardscreen.LeaderBoardDisplay;
-import com.deco2800.game.screens.leaderboardscreen.LeaderBoardScreen;
+import com.deco2800.game.screens.leaderboard.LeaderBoardDisplay;
+import com.deco2800.game.screens.leaderboard.LeaderBoardScreen;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
