+++ source/core/src/test/com/deco2800/game/screens/TitleScreenActionsTest.java
@@ -3,8 +3,7 @@ package com.deco2800.game.screens;
 import com.deco2800.game.GdxGame;
 import com.deco2800.game.extensions.GameExtension;
 import com.deco2800.game.generic.ServiceLocator;
-import com.deco2800.game.screens.contextscreen.ContextScreenActions;
-import com.deco2800.game.screens.titlescreen.TitleScreenActions;
+import com.deco2800.game.screens.title.TitleScreenActions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
